api/pyproject.toml
[build-system]
requires = ["poetry-core", "setuptools", "pip"]

[tool.poetry]
name = "reciper"
version = "0.1.0"
description = ""
authors = []
packages = [{include = "reciper"}]

[tool.poetry.dependencies]
python = "^3.11"
Celery = "*"
Django = "*"
django-environ = "*"
django-filter = "*"
djangorestframework = "*"
drf-spectacular = "*"
gunicorn = "*"
psycopg = { version = "*", extras = ["binary", "pool"] }
redis = "*"
sentry-sdk = "*"
django-storages = "*"
boto3 = "^1.26"
pip = "^24.0"
install = "^1.3.5"
djangorestframework-simplejwt = "^5.3.1"
pillow = "^10.3.0"

[tool.poetry.group.dev.dependencies]
Pygments = "*"
bandit = "*"
black = "*"
ruff = "*"
coverage = { version = "*", extras = ["toml"] }
django-debug-toolbar = "*"
django-extensions = "*"
Werkzeug = "*"
ipdb = "*"
ipython = "*"
model-bakery = "*"
pytest-django = "*"
pytest-mock = "*"
pytest = "*"
safety = "*"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "reciper.settings"
python_files = ["tests.py", "test_*.py", "*_test.py"]

[tool.black]
line-length = 120
target-version = ["py311",]
exclude = '''
/(
    \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.env
  | \.venv
  | _build
  | buck-out
  | build
  | dist
  | node_modules
)/
'''

[tool.ruff]
# Enable pycodestyle (`E`), Pyflakes (`F`), Isort (`I`), Bandit (`S`) and flake8 Django (`DJ`) codes.
select = ["E", "F", "I", "S", "DJ"]
ignore = []

# Allow autofix for all enabled rules (when `--fix`) is provided.
fixable = ["A", "B", "C", "D", "E", "F", "G", "I", "N", "Q", "S", "T", "W", "ANN", "ARG", "BLE", "COM", "DJ", "DTZ", "EM", "ERA", "EXE", "FBT", "ICN", "INP", "ISC", "NPY", "PD", "PGH", "PIE", "PL", "PT", "PTH", "PYI", "RET", "RSE", "RUF", "SIM", "SLF", "TCH", "TID", "TRY", "UP", "YTT"]
unfixable = []

# Exclude a variety of commonly ignored directories.
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
]

# Same as Black.
line-length = 120

# Allow unused variables when underscore-prefixed.
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"
target-version = "py311"

[tool.ruff.mccabe]
# Unlike Flake8, default to a complexity level of 10.
max-complexity = 10

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401", "F403"]
"conftest*" = ["F403"]
"test_*.py" = ["S"]
"tests.py" = ["S"]
"*_test.py" = ["S"]

[tool.ruff.isort]
lines-between-types = 1
lines-after-imports = 2
known-first-party=['reciper']
section-order= ['future', 'standard-library', 'third-party', 'django', 'first-party', 'local-folder']

[tool.ruff.isort.sections]
"django" = ["django"]

[tool.coverage.run]
omit = [
    "*migrations*",
    "*tests*",
    "/usr/*",
    "*/env/*",
    "*/urls.py",
    "reciper/fixtures/*",
    "reciper/apps/accounts/api/authentication.py",
    "reciper/apps/common/management/commands/generate_secretkey.py",
    "reciper/apps/common/management/commands/startapp.py",
    "reciper/apps/common/utils/redis.py",
    "reciper/loggers.py",
    "reciper/wsgi.py",
    "*/admin.py",
]
[tool.coverage.report]
exclude_lines =[
    "noqa",
    "pragma: no cover",
    "def __unicode__",
    "def __repr__",
    "def __str__",
    "if settings.DEBUG",
    "raise AssertionError",
    "raise NotImplementedError",
    "assert",
]
show_missing = true

